osa4, tehtävät 4.15 - 4.23

+++ osa4/blogilista/controllers/blogs.js
@@ -1,25 +1,49 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const { userExtractor } = require('../utils/middleware')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
-blogRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  if(blog.title && blog.url)
+blogRouter.post('/', userExtractor, async (request, response) => {
+  const body = request.body
+  if(body.title && body.url)
   {
+    const user = request.user
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes,
+      user: user._id
+    })
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog._id)
+
+    await user.save()
     response.status(201).json(savedBlog)
   }
-  else{
+  else
+  {
+    const blog = new Blog(body)
     response.status(400).json(blog)
   }
+
 })
-blogRouter.delete('/:id', async(request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+blogRouter.delete('/:id',userExtractor, async(request, response) => {
+  const user = request.user
+  const blog = await Blog.findById(request.params.id)
+  if(blog.user && blog.user.toString() === user._id.toString())
+  {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  }
+  else
+  {
+    response.status(400).json(blog)
+  }
 })
 blogRouter.put('/:id', async (request, response) => {
   const body = request.body
